Add fixture login in tests
